Added Utils namespace and reorganized
