+++ Core/ColumnLifts.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 using TeklaResultsInterrogator.Core;
+using TeklaResultsInterrogator.Utils;
 using TSD.API.Remoting.Loading;
 using TSD.API.Remoting.Solver;
 using TSD.API.Remoting.Structure;
@@ -85,22 +86,5 @@ namespace TeklaResultsInterrogator.Core
     {
         Bottom,
         Top,
-    }
-
-    public class NamedList<T>
-    {
-        public string Name { get; set; }
-        public List<T> Values { get; set; }
-
-        public NamedList(string name)
-        {
-            Name = name;
-            Values = new List<T>();
-        }
-
-        public void Add(T value)
-        {
-            Values.Add(value);
-        }
     }
 }
