+++ Utils/TextColor.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TeklaResultsInterrogator.Utils
+{
+    public enum TextColor
+    {
+        Text = ConsoleColor.White,
+        Command = ConsoleColor.Green,
+        Title = ConsoleColor.DarkCyan,
+        Path = ConsoleColor.DarkYellow,
+        Error = ConsoleColor.DarkRed,
+        Warning = ConsoleColor.Yellow,
+    }
+}
